Fixed Location Issues

+++ EcoScoop/app/src/main/java/edu/vassar/cmpu203/ecoscoop/src/view/DashboardFragment.java
@@ -86,29 +86,29 @@ public class DashboardFragment extends Fragment implements DashboardUI {
             pendingRetriever = retriever;
             return;
         }
-        binding.textCurrentTemp.setText(retriever.getCurrentTemp() + "°C");
-        binding.textCurrentWind.setText(retriever.getCurrentWind() + " km/h");
-        binding.textPrecipitation.setText(retriever.getDailyPrecip()[0] + " mm");
-        binding.textHumidity.setText(retriever.getHourlyHumidity()[0] + "%");
+        binding.textCurrentTemp.setText((int) retriever.getCurrentTemp() + "°C");
+        binding.textCurrentWind.setText((int)retriever.getCurrentWind() + " km/h");
+        binding.textPrecipitation.setText((int)retriever.getDailyPrecip()[0] + " mm");
+        binding.textHumidity.setText((int) retriever.getHourlyHumidity()[0] + "%");
 
         binding.cardTemperature.setOnClickListener(v ->
                 showPopup("Temperature Forecast",
-                        "High: " + retriever.getDailyTempMax()[0] + "°C\n" +
-                                "Low:  " + retriever.getDailyTempMin()[0] + "°C"));
+                        "High: " + (int)retriever.getDailyTempMax()[0] + "°C\n" +
+                                "Low:  " + (int)retriever.getDailyTempMin()[0] + "°C"));
 
         binding.cardWind.setOnClickListener(v ->
                 showPopup("Wind Detail",
-                        "Current: "   + retriever.getCurrentWind()    + " km/h\n" +
-                                "Max today: " + retriever.getDailyWindMax()[0] + " km/h"));
+                        "Current: "   + (int)retriever.getCurrentWind()    + " km/h\n" +
+                                "Max today: " + (int)retriever.getDailyWindMax()[0] + " km/h"));
 
         binding.cardPrecipitation.setOnClickListener(v ->
                 showPopup("Precipitation",
-                        "Today: "    + retriever.getDailyPrecip()[0] + " mm\n" +
-                                "Tomorrow: " + retriever.getDailyPrecip()[1] + " mm"));
+                        "Today: "    + (int) retriever.getDailyPrecip()[0] + " mm\n" +
+                                "Tomorrow: " + (int) retriever.getDailyPrecip()[1] + " mm"));
 
         binding.cardHumidity.setOnClickListener(v ->
                 showPopup("Humidity",
-                        "Current: " + retriever.getHourlyHumidity()[0] + "%"));
+                        "Current: " + (int) retriever.getHourlyHumidity()[0] + "%"));
     }
 
     private void showPopup(String title, String message) {
